+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
-## 4.26.0.pre
+## 4.25.4
 
-Released: [Place date of release here]
+Released: 2026-03-04
 
 ### New Features